get version for dcp_saf, move components to own file
